Add Dockerfile and other related files to enable support for containerization
This commit adds `Dockerfile`, `docker-compose.yaml` and `.dockerignore` for `gitdir`,
allowing the project to be deployed as a container.

I also embedded a small script called `gitdir_config` in it to allow first-time users
to initialize their Dockered `gitdir` more easily. See `docker-compose.yaml` for detail.

Note: This PR is experimental, if you're interested, please give it a try and send
      feedback. My eventual intention is to get this PR merged of course :)

Thanks for the nice project!